Update diary summary model to use whole number of total caloreis

+++ NutritionDiary.WebAPI/Models/DiarySummaryModel.cs
@@ -5,6 +5,6 @@ namespace NutritionDiary.WebAPI.Models
     public class DiarySummaryModel
     {
         public DateTime DiaryDate { get; set; }
-        public double TotalCalories { get; set; }
+        public int TotalCalories { get; set; }
     }
 }
